rename : ExerciseMethod 폴더 이동
